fix(update): never try to auto-play on idle tick (#23)
hopefully fix #23:
not able to reproduce consistently.

There are very few places the `play` command is sent:
- Select on queue item
- When START_PLAYING (autoplay) is in the update queue, which is
triggered when adding an artist, item to the queue, or on an idle
tick. This removes that last trigger which is not necessary.

Given that even when START_PLAYING is triggered, we always first ensure the
queue is not empty, I'm still not sure how this ever happens.